# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple REST API for shortening URLs.",
        "title": "URL Shortener API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/urls": {
            "post": {
                "description": "Generate a short link from the original URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URLs"
                ],
                "summary": "Create a shortened URL",
                "parameters": [
                    {
                        "description": "Original URL",
                        "name": "url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created",
                        "schema": {
                            "$ref": "#/definitions/github_com_zen-flo_url-shortener_internal_model.URL"
                        }
                    },
                    "400": {
                        "description": "invalid JSON",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/urls/{short}": {
            "get": {
                "description": "Retrieve the original URL by short code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URLs"
                ],
                "summary": "Get original URL",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"abc123\"",
                        "description": "Short code",
                        "name": "short",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Original URL retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_zen-flo_url-shortener_internal_model.URL"
                        }
                    },
                    "404": {
                        "description": "URL not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a short URL by its code",
                "tags": [
                    "URLs"
                ],
                "summary": "Delete a shortened URL",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"abc123\"",
                        "description": "Short code",
                        "name": "short",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to delete URL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_zen-flo_url-shortener_internal_model.URL": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "Timestamp when URL was created",
                    "type": "string"
                },
                "id": {
                    "description": "Unique identifier",
                    "type": "integer"
                },
                "original": {
                    "description": "Original URL",
                    "type": "string"
                },
                "short": {
                    "description": "Shortened URL",
                    "type": "string"
                }
            }
        }
    }
}